--- src/prop-validation/map-prop-types.js
@@ -1,14 +1,0 @@
-import PropTypes from "prop-types";
-
-export const mapPropTypes = PropTypes.shape({
-  center: PropTypes.shape({
-    longitude: PropTypes.number.isRequired,
-    latitude: PropTypes.number.isRequired,
-    zoom: PropTypes.number.isRequired,
-  }).isRequired,
-  options: PropTypes.arrayOf(PropTypes.shape({
-    id: PropTypes.string.isRequired,
-    longitude: PropTypes.number.isRequired,
-    latitude: PropTypes.number.isRequired
-  })),
-});
